Iteration 9 complete

+++ db/migrate/20180628203557_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.text :title
       t.text :description
-      t.float :price
+      t.integer :price
       t.text :image
 
       t.timestamps null: false
